+++ src/main/java/net/frozenblock/wilderwild/mixin/block/ice/HalfTransparentBlockMixin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024-2025 FrozenBlock
+ * Copyright 2023-2025 FrozenBlock
  * This file is part of Wilder Wild.
  *
  * This program is free software; you can redistribute it and/or
@@ -16,31 +16,34 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.mixin.snowlogging;
+package net.frozenblock.wilderwild.mixin.block.ice;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
+import net.frozenblock.wilderwild.registry.WWBlocks;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.HalfTransparentBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-@Mixin(Block.class)
-public class BlockMixin {
+@Mixin(HalfTransparentBlock.class)
+public class HalfTransparentBlockMixin {
 
 	@WrapOperation(
-		method = "spawnDestroyParticles",
+		method = "skipRendering",
 		at = @At(
 			value = "INVOKE",
-			target = "Lnet/minecraft/world/level/block/Block;getId(Lnet/minecraft/world/level/block/state/BlockState;)I"
+			target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z"
 		)
 	)
-	public int wilderWild$spawnDestroyParticles(BlockState state, Operation<Integer> original) {
-		if (SnowloggingUtils.isSnowlogged(state)) {
-			return original.call(SnowloggingUtils.getSnowEquivalent(state));
-		}
-		return original.call(state);
+	private boolean wilderWild$checkForIceAndFragileIce(
+		BlockState instance, Block block, Operation<Boolean> original
+	) {
+		if (instance.is(Blocks.ICE) && block == WWBlocks.FRAGILE_ICE) return true;
+		if (instance.is(WWBlocks.FRAGILE_ICE) && block == Blocks.ICE) return true;
+		return original.call(instance, block);
 	}
 
 }